:bulb: 마이페이지 컴포넌트 분리 2차
