+++ HikingClub/Feature/MyPage/MyPageViewController.swift
@@ -109,113 +109,3 @@ extension MyPageViewController: UITableViewDataSource {
         return tableHeaderView
     }
 }
-
-final class MyPageListHeaderButtonView: CodeBasedView {
-    enum ButtonType {
-        case myList
-        case savedList
-    }
-    
-    private var currentSelectedButton: ButtonType = .myList
-    
-    private let myListButton: UIButton = {
-        let button = MyPageListHeaderButton("나의 길")
-        return button
-    }()
-    
-    private let savedListButton: UIButton = {
-        let button = MyPageListHeaderButton("저장한 길")
-        return button
-    }()
-    
-    let currentSelctedRelay = PublishRelay<ButtonType>()
-    
-    private let disposeBag = DisposeBag()
-    
-    override func attribute() {
-        super.attribute()
-    }
-    
-    override func layout() {
-        addSubViews(myListButton, savedListButton)
-        myListButton.snp.makeConstraints {
-            $0.top.equalToSuperview()
-            $0.leading.equalToSuperview()
-            $0.bottom.equalToSuperview()
-        }
-        savedListButton.snp.makeConstraints {
-            $0.top.equalToSuperview()
-            $0.leading.equalTo(myListButton.snp.trailing).offset(20)
-            $0.trailing.equalToSuperview()
-            $0.bottom.equalToSuperview()
-        }
-    }
-    
-    override func bind() {
-        super.bind()
-        myListButton.rx.tap
-            .bind { [weak self] in self?.setSelected(.myList) }
-            .disposed(by: disposeBag)
-        
-        savedListButton.rx.tap
-            .bind { [weak self] in self?.setSelected(.savedList) }
-            .disposed(by: disposeBag)
-    }
-
-    func setSelected(_ type: ButtonType) {
-        switch type {
-        case .myList:
-            myListButton.isSelected = true
-            savedListButton.isSelected = false
-        case .savedList:
-            myListButton.isSelected = false
-            savedListButton.isSelected = true
-        }
-    }
-}
-
-final class MyPageListHeaderButton: UIButton, CodeBasedProtocol {
-    private let barIndicator = UIView()
-    private let contentLabel: UILabel = {
-        let label = UILabel()
-        label.setFont(.semiBold18)
-        return label
-    }()
-    
-    override var isSelected: Bool {
-        didSet {
-            updatedSelectStatus()
-        }
-    }
-    
-    init(_ title: String) {
-        super.init(frame: .zero)
-        contentLabel.text = title
-        barIndicator.isHidden = true
-        layout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-        
-    func layout() {
-        addSubViews(contentLabel, barIndicator)
-        contentLabel.snp.makeConstraints {
-            $0.top.equalToSuperview()
-            $0.leading.trailing.equalToSuperview()
-        }
-        barIndicator.snp.makeConstraints {
-            $0.top.equalTo(contentLabel.snp.bottom).offset(6)
-            $0.leading.trailing.equalToSuperview()
-            $0.bottom.equalToSuperview()
-            $0.height.equalTo(2)
-        }
-    }
-    
-    private func updatedSelectStatus() {
-        contentLabel.textColor = isSelected ? .green500 : .gray300
-        barIndicator.backgroundColor = isSelected ? .green500 : .clear
-        barIndicator.isHidden = !isSelected
-    }
-}
